fix(cli): simplify Cline credits error matcher
