+++ apps/cli/src/tui/cline-account.ts
@@ -14,6 +14,8 @@ import { toProviderApiKey } from "../utils/provider-auth";
 import type { Config } from "../utils/types";
 
 const WORKOS_TOKEN_PREFIX = "workos:";
+export const CLINE_CREDITS_DASHBOARD_URL =
+	"https://app.cline.bot/dashboard/account?tab=credits&redirect=true";
 
 type ClineAccountConfig = Pick<Config, "apiKey" | "providerId">;
 
@@ -35,6 +37,14 @@ export function isClineAccountAuthErrorMessage(message: string): boolean {
 	return (
 		normalized === "no cline account auth token found" ||
 		normalized.includes("requires re-authentication")
+	);
+}
+
+export function isClineCreditsBalanceErrorMessage(message: string): boolean {
+	const normalized = message.trim().toLowerCase();
+	return (
+		normalized.includes("insufficient balance") &&
+		normalized.includes("cline credits balance")
 	);
 }
